# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CAPPirote",
    "description": "Documentación de la API CAPPirote",
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "https://cappirote-bf5611b04e17.herokuapp.com/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/images/{id_hdad}": {
      "get": {
        "tags": [
          "imagenes"
        ],
        "parameters": [
          {
            "name": "id_hdad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "imagenes"
        ],
        "parameters": [
          {
            "name": "id_hdad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "image_urls": {
                    "type": "object",
                    "example": "any"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "409": {
            "description": "Conflict",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/images/{id_hdad}/{image_url}": {
      "delete": {
        "tags": [
          "imagenes"
        ],
        "summary": "Eliminar una imagen asociada a una hermandad",
        "description": "Elimina la imagen asociada a la hermandad con el `id_hdad` y el `image_url` proporcionados.",
        "operationId": "deleteImage",
        "parameters": [
          {
            "name": "id_hdad",
            "in": "path",
            "description": "ID de la hermandad que contiene la imagen.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "image_url",
            "in": "path",
            "description": "URL de la imagen a eliminar.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token JWT para autenticación (si es necesario).",
            "schema": {
              "type": "string",
              "default": "Bearer {token}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen eliminada con éxito.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Imagen borrada con éxito"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta (falta `id_hdad` o `image_url`, o el formato de `id_hdad` es incorrecto).",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Faltan parámetros necesarios: id_hdad o image_url"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Imagen no encontrada o no existe.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Imagen no encontrada o no existe"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor al eliminar la imagen.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Error eliminando la imagen"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/devotos/{id_hdad}/usuario/{id_usuario}": {
      "delete": {
        "tags": [
          "devotos"
        ],
        "summary": "Elimina la relación entre un usuario y una hermandad.",
        "description": "Elimina la relación entre un usuario y una hermandad.",
        "operationId": "deleteDevoto",
        "parameters": [
          {
            "name": "id_hdad",
            "in": "path",
            "description": "ID de la hermandad",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 4
            }
          },
          {
            "name": "id_usuario",
            "in": "path",
            "description": "ID del usuario",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devoto eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string",
                      "example": "Devoto eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Faltan parámetros en la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Faltan parámetros necesarios (id_hdad, id_usuario)."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró el devoto con los ID proporcionados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Devoto no encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno en el servidor al intentar eliminar el devoto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Error interno del servidor al eliminar el devoto."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
...
}